Added comments, cleaned up code

+++ bamazon.sql
@@ -8,7 +8,7 @@ CREATE TABLE products (
 	item_id INT NOT NULL,
 	product_name VARCHAR(50) NULL,
 	department_name VARCHAR(20) NOT NULL,
-	price INT(10) NOT NULL,
+	price DECIMAL(10,2) NOT NULL,
 	stock_quantity INT (5) NOT NULL,
 	PRIMARY KEY (item_id)
 );
